+++ GRDB Tests/Public/Record/PrimaryKeyRowIDTests.swift
@@ -500,13 +500,13 @@ class PrimaryKeyRowIDTests: GRDBTestCase {
                 
                 do {
                     let ids: [Int64] = []
-                    let fetchedRecords = Array(Person.fetch(db, primaryKeys: ids))
+                    let fetchedRecords = Array(Person.fetch(db, keys: ids))
                     XCTAssertEqual(fetchedRecords.count, 0)
                 }
                 
                 do {
                     let ids = [record1.id!, record2.id!]
-                    let fetchedRecords = Array(Person.fetch(db, primaryKeys: ids))
+                    let fetchedRecords = Array(Person.fetch(db, keys: ids))
                     XCTAssertEqual(fetchedRecords.count, 2)
                     XCTAssertEqual(Set(fetchedRecords.map { $0.id }), Set(ids))
                 }
@@ -524,13 +524,13 @@ class PrimaryKeyRowIDTests: GRDBTestCase {
                 
                 do {
                     let ids: [Int64] = []
-                    let fetchedRecords = Person.fetchAll(db, primaryKeys: ids)
+                    let fetchedRecords = Person.fetchAll(db, keys: ids)
                     XCTAssertEqual(fetchedRecords.count, 0)
                 }
                 
                 do {
                     let ids = [record1.id!, record2.id!]
-                    let fetchedRecords = Person.fetchAll(db, primaryKeys: ids)
+                    let fetchedRecords = Person.fetchAll(db, keys: ids)
                     XCTAssertEqual(fetchedRecords.count, 2)
                     XCTAssertEqual(Set(fetchedRecords.map { $0.id }), Set(ids))
                 }
@@ -546,12 +546,12 @@ class PrimaryKeyRowIDTests: GRDBTestCase {
                 
                 do {
                     let id: Int64? = nil
-                    let fetchedRecord = Person.fetchOne(db, primaryKey: id)
+                    let fetchedRecord = Person.fetchOne(db, key: id)
                     XCTAssertTrue(fetchedRecord == nil)
                 }
                 
                 do {
-                    let fetchedRecord = Person.fetchOne(db, primaryKey: record.id)!
+                    let fetchedRecord = Person.fetchOne(db, key: record.id)!
                     XCTAssertTrue(fetchedRecord.id == record.id)
                     XCTAssertTrue(fetchedRecord.name == record.name)
                     XCTAssertTrue(fetchedRecord.age == record.age)