+++ GRDB/Core/Protocols/DatabaseTableMapping.swift
@@ -4,7 +4,7 @@
 /// The protocol comes with built-in methods that allow to fetch instances
 /// identified by their primary key, or any other key:
 ///
-///     Person.fetchOne(db, primaryKey: 123)  // Person?
+///     Person.fetchOne(db, key: 123)  // Person?
 ///     Citizenship.fetchOne(db, key: ["personId": 12, "countryId": 45]) // Citizenship?
 ///
 /// DatabaseTableMapping is adopted by Record.
@@ -14,17 +14,19 @@ public protocol DatabaseTableMapping : RowConvertible {
 
 extension DatabaseTableMapping {
     
+    // MARK: - Single-Column Primary Key
+    
     /// Fetches a sequence of values, given their primary keys.
     ///
-    ///     let persons = Person.fetch(db, primaryKeys: [1, 2, 3]) // DatabaseSequence<Person>
+    ///     let persons = Person.fetch(db, keys: [1, 2, 3]) // DatabaseSequence<Person>
     ///
     /// The order of values in the returned sequence is undefined.
     ///
     /// - parameter db: A Database.
-    /// - parameter primaryKeys: An array of primary keys.
+    /// - parameter keys: An array of primary keys.
     /// - returns: A sequence.
-    public static func fetch<Sequence: SequenceType where Sequence.Generator.Element: DatabaseValueConvertible>(db: Database, primaryKeys: Sequence) -> DatabaseSequence<Self> {
-        if let statement = self.fetchByPrimaryKeyStatement(db, primaryKeys: primaryKeys) {
+    public static func fetch<Sequence: SequenceType where Sequence.Generator.Element: DatabaseValueConvertible>(db: Database, keys: Sequence) -> DatabaseSequence<Self> {
+        if let statement = self.fetchByPrimaryKeyStatement(db, keys: keys) {
             return self.fetch(statement)
         } else {
             return DatabaseSequence()
@@ -33,15 +35,15 @@ extension DatabaseTableMapping {
     
     /// Fetches an array of values, given their primary keys.
     ///
-    ///     let persons = Person.fetchAll(db, primaryKeys: [1, 2, 3]) // [Person]
+    ///     let persons = Person.fetchAll(db, keys: [1, 2, 3]) // [Person]
     ///
     /// The order of values in the returned array is undefined.
     ///
     /// - parameter db: A Database.
-    /// - parameter primaryKeys: An array of primary keys.
+    /// - parameter keys: An array of primary keys.
     /// - returns: An array.
-    public static func fetchAll<Sequence: SequenceType where Sequence.Generator.Element: DatabaseValueConvertible>(db: Database, primaryKeys: Sequence) -> [Self] {
-        if let statement = self.fetchByPrimaryKeyStatement(db, primaryKeys: primaryKeys) {
+    public static func fetchAll<Sequence: SequenceType where Sequence.Generator.Element: DatabaseValueConvertible>(db: Database, keys: Sequence) -> [Self] {
+        if let statement = self.fetchByPrimaryKeyStatement(db, keys: keys) {
             return self.fetchAll(statement)
         } else {
             return []
@@ -50,17 +52,62 @@ extension DatabaseTableMapping {
     
     /// Fetches a single value given its primary key.
     ///
-    ///     let person = Person.fetchOne(db, primaryKey: 123) // Person?
+    ///     let person = Person.fetchOne(db, key: 123) // Person?
     ///
     /// - parameter db: A Database.
-    /// - parameter primaryKey: A value.
+    /// - parameter key: A value.
     /// - returns: An optional value.
-    public static func fetchOne<PrimaryKeyType: DatabaseValueConvertible>(db: Database, primaryKey: PrimaryKeyType?) -> Self? {
-        guard let primaryKey = primaryKey else {
+    public static func fetchOne<PrimaryKeyType: DatabaseValueConvertible>(db: Database, key: PrimaryKeyType?) -> Self? {
+        guard let key = key else {
             return nil
         }
-        return self.fetchOne(self.fetchByPrimaryKeyStatement(db, primaryKeys: [primaryKey])!)
+        return self.fetchOne(self.fetchByPrimaryKeyStatement(db, keys: [key])!)
     }
+    
+    // Returns "SELECT * FROM table WHERE id IN (?,?,?)"
+    //
+    // Returns nil if keys is empty.
+    private static func fetchByPrimaryKeyStatement<Sequence: SequenceType where Sequence.Generator.Element: DatabaseValueConvertible>(db: Database, keys: Sequence) -> SelectStatement? {
+        // Fail early if databaseTable is nil
+        guard let databaseTableName = self.databaseTableName() else {
+            fatalError("Nil returned from \(self).databaseTableName()")
+        }
+        
+        // Fail early if database table does not exist.
+        guard let primaryKey = db.primaryKeyForTable(named: databaseTableName) else {
+            fatalError("Table \(databaseTableName.quotedDatabaseIdentifier) does not exist. See \(self).databaseTableName()")
+        }
+        
+        // Fail early if database table has not one column in its primary key
+        let columns = primaryKey.columns
+        guard columns.count == 1 else {
+            fatalError("Primary key of table \(databaseTableName.quotedDatabaseIdentifier) is not made of a single column. See \(self).databaseTableName()")
+        }
+        
+        let keys = keys.map { $0 as DatabaseValueConvertible? }
+        
+        switch keys.count {
+        case 0:
+            // Avoid performing useless SELECT
+            return nil
+        case 1:
+            // Use '=' in SQL query
+            let sql = "SELECT * FROM \(databaseTableName.quotedDatabaseIdentifier) WHERE \(columns.first!.quotedDatabaseIdentifier) = ?"
+            let statement = db.selectStatement(sql)
+            statement.arguments = StatementArguments(keys)
+            return statement
+        default:
+            // Use 'IN'
+            let questionMarks = Array(count: keys.count, repeatedValue: "?").joinWithSeparator(",")
+            let sql = "SELECT * FROM \(databaseTableName.quotedDatabaseIdentifier) WHERE \(columns.first!.quotedDatabaseIdentifier) IN (\(questionMarks))"
+            let statement = db.selectStatement(sql)
+            statement.arguments = StatementArguments(keys)
+            return statement
+        }
+    }
+    
+    
+    // MARK: - Other Keys
     
     /// Fetches a sequence of values, given an array of key dictionaries.
     ///
@@ -81,7 +128,7 @@ extension DatabaseTableMapping {
     
     /// Fetches an array of values, given an array of key dictionaries.
     ///
-    ///     let persons = Person.fetchAll(db, primaryKeys: [["name": "Arthur"], ["name": "Barbara"]]) // [Person]
+    ///     let persons = Person.fetchAll(db, keys: [["name": "Arthur"], ["name": "Barbara"]]) // [Person]
     ///
     /// The order of values in the returned array is undefined.
     ///
@@ -96,7 +143,7 @@ extension DatabaseTableMapping {
         }
     }
     
-    /// Fetches a single value given a key.
+    /// Fetches a single value given a key dictionary.
     ///
     ///     let person = Person.fetchOne(db, key: ["name": Arthur"]) // Person?
     ///
@@ -105,48 +152,6 @@ extension DatabaseTableMapping {
     /// - returns: An optional value.
     public static func fetchOne(db: Database, key: [String: DatabaseValueConvertible?]) -> Self? {
         return self.fetchOne(self.fetchByKeyStatement(db, keys: [key])!)
-    }
-    
-    // Returns "SELECT * FROM table WHERE id IN (?,?,?)"
-    //
-    // Returns nil if primaryKeys is empty.
-    private static func fetchByPrimaryKeyStatement<Sequence: SequenceType where Sequence.Generator.Element: DatabaseValueConvertible>(db: Database, primaryKeys: Sequence) -> SelectStatement? {
-        // Fail early if databaseTable is nil
-        guard let databaseTableName = self.databaseTableName() else {
-            fatalError("Nil returned from \(self).databaseTableName()")
-        }
-        
-        // Fail early if database table does not exist.
-        guard let primaryKey = db.primaryKeyForTable(named: databaseTableName) else {
-            fatalError("Table \(databaseTableName.quotedDatabaseIdentifier) does not exist. See \(self).databaseTableName()")
-        }
-        
-        // Fail early if database table has not one column in its primary key
-        let columns = primaryKey.columns
-        guard columns.count == 1 else {
-            fatalError("Primary key of table \(databaseTableName.quotedDatabaseIdentifier) is not made of a single column. See \(self).databaseTableName()")
-        }
-        
-        let primaryKeys = primaryKeys.map { $0 as DatabaseValueConvertible? }
-        
-        switch primaryKeys.count {
-        case 0:
-            // Avoid performing useless SELECT
-            return nil
-        case 1:
-            // Use '=' in SQL query
-            let sql = "SELECT * FROM \(databaseTableName.quotedDatabaseIdentifier) WHERE \(columns.first!.quotedDatabaseIdentifier) = ?"
-            let statement = db.selectStatement(sql)
-            statement.arguments = StatementArguments(primaryKeys)
-            return statement
-        default:
-            // Use 'IN'
-            let questionMarks = Array(count: primaryKeys.count, repeatedValue: "?").joinWithSeparator(",")
-            let sql = "SELECT * FROM \(databaseTableName.quotedDatabaseIdentifier) WHERE \(columns.first!.quotedDatabaseIdentifier) IN (\(questionMarks))"
-            let statement = db.selectStatement(sql)
-            statement.arguments = StatementArguments(primaryKeys)
-            return statement
-        }
     }
     
     // Returns "SELECT * FROM table WHERE (a = ? AND b = ?) OR (a = ? AND b = ?) ...